Add files via upload
Here lies the solution to the HackerRank problem: https://www.hackerrank.com/challenges/matrix-rotation-algo/problem. The problem statement is: You are given a 2D matrix, a, of dimension MxN and a positive integer R. You have to rotate the matrix R times and print the resultant matrix. Rotation should be in an anti-clockwise direction.